+++ src/preprocess/base.py
@@ -1,7 +1,9 @@
 from pathlib import Path
-from collections import namedtuple
 
-Region = namedtuple('Region', ['name', 'lonmin', 'lonmax', 'latmin', 'latmax'])
+from ..utils import Region, get_kenya
+
+
+__all__ = ['BasePreProcessor', 'Region', 'get_kenya']
 
 
 class BasePreProcessor:
@@ -19,12 +21,3 @@ class BasePreProcessor:
 
         if not self.interim_folder.exists():
             self.interim_folder.mkdir()
-
-    @staticmethod
-    def get_kenya() -> Region:
-        """This pipeline is focused on drought prediction in Kenya.
-        This function allows Kenya's bounding box to be easily accessed
-        by all exporters.
-        """
-        return Region(name='kenya', lonmin=33.501, lonmax=42.283,
-                      latmin=-5.202, latmax=6.002)
